Add instance filtering to declaration and refactor subformula variables
Declaration can now contain a formula that filters in advance matching
component instances.
Note that this feature is meant to be used EXCLUSIVELY with disjunctive
style. By moving guard predicates in the filter it is easy to encounter
scenarios where the set of actual instances matched by the declaration
is empty, which for existential rules leads to contradictions that
wouldn't occur had we embedded the same predicates into a conjunctive
term (which could still yield "true" if the premise is "false").

All subformulas are now stored in sets, so that conjunction/disjunction
of equal terms is automatically ignored.

More code/packages refactoring.