Add DReAM AbstractComponent and behavior package
Add AbstractComponent implementing base DReAM components and classes to
model their behaviors (LTS made of Transition objects connecting
ControlLocation instances).
TriggerTransition Operation implements the effect of executing the
transition on the LTS, and is instanced on-demand from the current LTS
state and the selected Interaction.
Additionally, add the Idle Formula.